Enable uploading and downloading
When uploading,
If we found a client save with the same game id, we copy the player
files into it from the host save.
If we didn't find it, we copy the whole host save over to the client
directory.

When downloading, copy the player files from client if they exist into
the host save.