Change pub/sub messaging methodology to async/await
For messages back and forth between the renderer and main, it used to
use the pub/sub way of doing things via messages. Changed this to
async/await, which is much cleaner and is just more "normal" and
understandable.

I also made a custom hook to implement the state of the requests to the
main, taking inspiration from React Query. Note that I didn't actually
use React Query because the renderer isn't actually doing a network
request to the main, its just sending and receiving a message, so we
only really needed to keep track of certain states.